use an accordion in the sidebar
There are 3 items in the sidebar (folders, outline, links).
Previously links were put under the outline tab but that
does not work well, particularly when the outline is larger then the screen.

Outline and links show up contextually for documents that have them.
We may include other contextual accordion items in the future.

Other fixes
* links from comments are prettified and shown in the links accordion item
* fix window scroll when viewing an individual comment
* fix avoid calling the export API multiple times
* get rid of setTimeout hacks
* check for "_Reassigned to" in comments.
